refactor: created FallBack component to display FallBack text when there is no Meals filters or when there is no Favorites
