+++ components/DefaultText.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { Text, StyleSheet } from 'react-native';
 
 const DefaultText = (props)=>{
     return <Text style={{...styles.text, ...props.style}}>{props.children}</Text>
